requirements.txt
langchain_openai 
langchain_core
streamlit
python-dotenv